test: AdminThemeController 테스트 추가
