+++ src/test/java/roomescape/admin/presentation/AdminViewControllerTest.java
@@ -1,4 +1,4 @@
-package roomescape.view;
+package roomescape.admin.presentation;
 
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
@@ -15,7 +15,6 @@ import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.web.servlet.MockMvc;
 import roomescape.admin.AdminHandlerInterceptor;
-import roomescape.admin.presentation.AdminViewController;
 import roomescape.login.LoginMemberArgumentResolver;
 
 @WebMvcTest(AdminViewController.class)
