+++ src/test/java/roomescape/theme/presentation/ThemeControllerTest.java
@@ -1,11 +1,8 @@
 package roomescape.theme.presentation;
 
 import static org.mockito.Mockito.when;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.header;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
@@ -20,8 +17,6 @@ import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 import roomescape.admin.AdminHandlerInterceptor;
 import roomescape.login.LoginMemberArgumentResolver;
-import roomescape.theme.dto.ThemeAddRequest;
-import roomescape.theme.dto.ThemeResponse;
 import roomescape.theme.service.ThemeService;
 
 @WebMvcTest(ThemeController.class)
@@ -53,21 +48,6 @@ class ThemeControllerTest {
                 .andExpect(jsonPath("$").isArray());
     }
 
-    @DisplayName("테마 추가 요청을 처리할 수 있다")
-    @Test
-    void should_handle_post_theme_request_when_requested() throws Exception {
-        ThemeAddRequest themeAddRequest = new ThemeAddRequest("리비 테마", "리비 설명", "리비 경로");
-        ThemeResponse mockResponse = new ThemeResponse(1L, "리비 테마", "리비 설명", "리비 경로");
-
-        when(themeService.saveTheme(themeAddRequest)).thenReturn(mockResponse);
-
-        mockMvc.perform(post("/themes")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(objectMapper.writeValueAsString(themeAddRequest)))
-                .andExpect(status().isCreated())
-                .andExpect(header().string("Location", "/themes/" + mockResponse.id()));
-    }
-
     @DisplayName("인기 테마 목록을 읽는 요청을 처리할 수 있다")
     @Test
     void should_handle_get_popular_themes_request_when_requested() throws Exception {
@@ -77,12 +57,5 @@ class ThemeControllerTest {
                 .andExpect(status().isOk())
                 .andExpect(content().contentType(MediaType.APPLICATION_JSON))
                 .andExpect(jsonPath("$").isArray());
-    }
-
-    @DisplayName("테마 삭제 요청을 처리할 수 있다")
-    @Test
-    void should_handle_delete_theme_when_requested() throws Exception {
-        mockMvc.perform(delete("/themes/{id}", 1))
-                .andExpect(status().isNoContent());
     }
 }
